moved color and style to StyleHelper
