+++ Screens/AddDiet.js
@@ -3,6 +3,7 @@ import { View, Text, TextInput, Alert, StyleSheet, TouchableOpacity, Button } fr
 import DateTimePicker from '@react-native-community/datetimepicker';
 import { DataContext } from '../context/DataContext';
 import { ThemeContext } from '../context/ThemeContext';
+import { styles } from '../style/StyleHelper';
 
 export default function AddDiet({ navigation }) {
     const { themeStyles } = useContext(ThemeContext);
@@ -52,7 +53,7 @@ export default function AddDiet({ navigation }) {
     }
 
     return (
-        <View style={styles.container}>
+        <View style={[styles.addScreenContainer, { backgroundColor: themeStyles.backgroundColor }]}>
             <Text style={[styles.label, { color: themeStyles.textColor }]}>Description *</Text>
             <TextInput
                 style={[styles.input, styles.tallInput]}
@@ -90,43 +91,3 @@ export default function AddDiet({ navigation }) {
         </View>
     );
 }
-
-const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        padding: 10,
-        backgroundColor: '#a3b18a',
-    },
-    label: {
-        marginBottom: 5,
-        fontSize: 14,
-        color: '#344e41',
-    },
-    input: {
-        borderWidth: 1,
-        borderColor: '#3a5a40',
-        padding: 8,
-        marginBottom: 10,
-        borderRadius: 5,
-        backgroundColor: '#fff',
-        justifyContent: 'center',
-    },
-    tallInput: {
-        height: 100,
-    },
-    buttonContainer: {
-      flexDirection: 'row',
-      justifyContent: 'space-between',
-      marginTop: 10,
-    },
-    button: {
-      backgroundColor: '#3a5a40',
-      padding: 8,
-      borderRadius: 5,
-      flex: 0.48,
-    },
-    buttonText: {
-      color: '#000',
-      textAlign: 'center',
-    },
-});
