+++ Screens/Diet.js
@@ -1,10 +1,11 @@
 import React, { useContext } from 'react';
-import { View, Text, TouchableOpacity, StyleSheet } from 'react-native';
+import { View, Text, TouchableOpacity } from 'react-native';
 import ItemsList from '../components/ItemsList';
-import { ThemeContext } from '../context/ThemeContext'; 
+import { ThemeContext } from '../context/ThemeContext';
+import { styles } from '../style/StyleHelper';  
 
 export default function Diet({ navigation }) {
-  const { themeStyles } = useContext(ThemeContext);  
+  const { themeStyles } = useContext(ThemeContext);
 
   React.useLayoutEffect(() => {
     navigation.setOptions({
@@ -13,11 +14,11 @@ export default function Diet({ navigation }) {
           onPress={() => navigation.navigate('AddDiet')}
           style={[styles.addButton, { backgroundColor: themeStyles.backgroundColor }]}  
         >
-          <Text style={{ color: themeStyles.textColor }}>Add</Text> 
+          <Text style={{ color: themeStyles.textColor }}>Add</Text>
         </TouchableOpacity>
       ),
     });
-  }, [navigation, themeStyles]);  
+  }, [navigation, themeStyles]);
 
   return (
     <View style={[styles.container, { backgroundColor: themeStyles.backgroundColor }]}>
@@ -25,16 +26,3 @@ export default function Diet({ navigation }) {
     </View>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1, 
-    justifyContent: 'center',
-    alignItems: 'stretch', 
-    padding: 10, 
-  },
-  addButton: {
-    padding: 10,
-    borderRadius: 5, 
-  },
-});
